+++ Project/ZPGProject/ZPGProject/Scene/SceneC5_4Objects/SceneC5_4Objects.h
@@ -27,7 +27,7 @@ protected:
 public:
 	SceneC5_4Objects(GLFWwindow* window);
 	using Scene::tick;
-	void tick(double deltaTime);
+	void tick(float deltaTime);
 	using Scene::draw;
 	void draw();
 };
